delete mounts module
