+++ pvlib/pvsystem.py
@@ -190,9 +190,8 @@ class PVSystem:
                  racking_model=None, losses_parameters=None, name=None):
 
         if arrays is None:
-            from pvlib import mounts
             self.arrays = (Array(
-                mounts.FixedMount(surface_tilt, surface_azimuth),
+                FixedMount(surface_tilt, surface_azimuth),
                 albedo,
                 surface_type,
                 module,
@@ -1162,8 +1161,7 @@ class Array:
                  modules_per_string=1, strings=1,
                  racking_model=None, name=None):
         if mount is None:
-            from pvlib import mounts  # avoid circular import issue
-            self.mount = mounts.FixedMount(0, 180)
+            self.mount = FixedMount(0, 180)
         else:
             self.mount = mount
 
@@ -1379,6 +1377,54 @@ class Array:
                              'option for Array')
         else:
             raise ValueError(model + ' is not a valid IAM model')
+
+
+class FixedMount:
+    def __init__(self, surface_tilt=0, surface_azimuth=180):
+        self.surface_tilt = surface_tilt
+        self.surface_azimuth = surface_azimuth
+
+    def __repr__(self):
+        return (
+            'FixedMount:'
+            f'\n    surface_tilt: {self.surface_tilt}'
+            f'\n    surface_azimuth: {self.surface_azimuth}'
+        )
+
+    def get_orientation(self, solar_zenith, solar_azimuth):
+        return {
+            'surface_tilt': self.surface_tilt,
+            'surface_azimuth': self.surface_azimuth,
+        }
+
+
+class SingleAxisTrackerMount:
+    def __init__(self, axis_tilt, axis_azimuth, max_angle, backtrack, gcr,
+                 cross_axis_tilt):
+        self.axis_tilt = axis_tilt
+        self.axis_azimuth = axis_azimuth
+        self.max_angle = max_angle
+        self.backtrack = backtrack
+        self.gcr = gcr
+        self.cross_axis_tilt = cross_axis_tilt
+
+    def __repr__(self):
+        attrs = ['axis_tilt', 'axis_azimuth', 'max_angle',
+                 'backtrack', 'gcr', 'cross_axis_tilt']
+
+        return 'SingleAxisTrackerMount:\n    ' + '\n    '.join(
+            f'{attr}: {getattr(self, attr)}' for attr in attrs
+        )
+
+    def get_orientation(self, solar_zenith, solar_azimuth):
+        from pvlib import tracking  # avoid circular import issue
+        tracking_data = tracking.singleaxis(
+            solar_zenith, solar_azimuth,
+            self.axis_tilt, self.axis_azimuth,
+            self.max_angle, self.backtrack,
+            self.gcr, self.cross_axis_tilt
+        )
+        return tracking_data
 
 
 def calcparams_desoto(effective_irradiance, temp_cell,
